fix(数据源): 切换导航主菜单时，闪现其他的空状态图片
