+++ core/core-frontend/src/views/visualized/data/datasource/form/ApiHttpRequestDraw.vue
@@ -11,7 +11,7 @@ import EmptyBackground from '@/components/empty-background/src/EmptyBackground.v
 import { checkApiItem } from '@/api/datasource'
 import { cloneDeep } from 'lodash-es'
 import { fieldType } from '@/utils/attr'
-import { ApiConfiguration } from '@/views/visualized/data/datasource/form/index.vue'
+import type { ApiConfiguration } from '@/views/visualized/data/datasource/form/option'
 
 export interface Field {
   name: string
